decaf377_rdsa
Trait Domain

pub trait Domain: Sealed { }

Abstracts over different signature domains.

This design is described at the end of 5.4.6 of the Zcash protocol specification: the generator used for the signature scheme is left as an unspecified parameter, chosen differently for each signature domain.

To handle this, we encode the domain as a type parameter.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.